requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector[pymysql]==1.11.0
cryptography==43.0.0
dataclasses-json==0.6.7
deepgram-sdk==3.4.0
deprecation==2.1.0
docker==7.1.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google==3.0.0
google-auth==2.32.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numpy==2.0.0
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydub==0.25.1
PyJWT==2.8.0
PyMySQL==1.1.1
pyngrok==7.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==2.0.31
twilio==9.2.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4